web_scraper_files/requirements.txt
python-dotenv
langdetect
psycopg[binary]
requests
beautifulsoup4
trafilatura
readability-lxml
feedparser
tf-playwright-stealth
pgvector
sentence-transformers
